1. User 평점 타입 변경 (Double -> Integer) 및 반올림 저장 로직 추가 2. application.yml JPA 설정 변경 (ddl-auto: validate, show-sql 제거) 3. ContactDto Schema 어노테이션 제거

+++ backend/src/main/java/com/jichan/entity/User.java
@@ -49,7 +49,7 @@ public class User extends BaseTimeEntity {
 
     @Builder.Default
     @Column(name = "average_rating", nullable = false)
-    private double averageRating = 0.0; // Changed to double
+    private int averageRating = 0; // Changed to double
 
     @Builder.Default
     @Column(name = "review_count", nullable = false)
@@ -75,7 +75,7 @@ public class User extends BaseTimeEntity {
         this.emailVerified = true;
     }
 
-    public void updateRating(double averageRating) {
+    public void updateRating(int averageRating) {
         this.averageRating = averageRating;
     }
